Committing GUI Mockups
>GoUI.html
Basic UI with alternating black and white turn based on the location of
the click event.  Opens a window alert to show the position, in terms of
the board,  to show position to be placed in.

> basicCanvas(13x13) etc
13x13, 9x9 and 19x19 boards. Place alternating pieces at nearest point
to user click

> jcanvas.min.js, jquery-ui.js, jquery.js, json2.js
JavaScript Libraries which make Javascript nice. Need to be stored in
the same directory as the files for them to run.